refactor(server): Refactor TemplateService to use FileSystem
Inject the FileSystem trait into TemplateService to decouple it from the concrete file system. This allows for easier testing by mocking file system interactions. Update the instantiation of TemplateService in ConfigServerApp to provide the DefaultFileSystem. Add a new test suite for TemplateService with a test case that verifies the refactoring.