feat(cert): Use -networks flag instead of -ip for signing
Refactors the certificate signing process to use the `-networks` flag,
which accepts a list of CIDR networks, instead of the `-ip` flag.

This is a breaking change to the configuration. The `ip: String` field
in `HostConfig` has been replaced with `networks: List[String]`.

The `nebula-cert sign` command is now constructed with the `-networks`
flag, and a symbolic link is created for each network specified in the
host's configuration, pointing to the generated certificate.